build: Fix tests/clone.scm failure when using a source tarball
Test tests/clone.scm uses sshd when it is available, but files under
tests/.ssh directory, which are required to run sshd, are not included
in the source tarball. Therefore, the test fails with chmod failure
because files don't exist. To fix it, add them to EXTRA_DIST.

This problem was reported by an Arch Linux user on AUR website:
https://aur.archlinux.org/packages/guile-git-lib/#comment-749163