Add sanitycheck utility to ensure libgit2 compatibility
Upstream is not very active at the moment and not always up-to-date
regarding libgit2 releases. Occasionally, the definition for struct
layouts (used for the Guile FFI) therefore mismatch those of libgit2.
This can cause extremely weird and hard to debug errors, therefore
this patch adds a Scheme script which ensures that the size of these
Guile struct layouts matches the size of the libgit2 C structs as
determined via sizeof().

The checks can be invoked using `make sanitycheck`.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>